wip: add l [conv↑] l assumption to Lift-η
This is an attempt to make the proof of full reduction go through.

I had to add convertibility assumptions to Definition.Conversion.Lift.
I am now stuck in two places:

- In lift~toConv↓′ for strong Σ, the second recursive call seems to need
  a proof that B [ _ ]₀ is convertible to itself, but we only know that
  B is convertible to itself.
- In ~-to-conv, we need B to be convertible to itself but we only know
  that it is definitionally equal to A.

I wonder if we should have an admissible rule that Γ ⊢ t [conv↑] u ∷ A
implies Γ ⊢ A [conv↑] A (and similarly for the other term relations).